+++ pyinterpolate/variogram/empirical/experimental_variogram.py
@@ -1,5 +1,5 @@
 """
-The main class for experimental semivariogram.
+Class for Experimental Variogram and helper class for DirectionalVariograms.
 
 Authors
 -------
@@ -15,6 +15,149 @@ from prettytable import PrettyTable
 from pyinterpolate.variogram.empirical.covariance import calculate_covariance
 from pyinterpolate.variogram.empirical.semivariance import calculate_semivariance
 from pyinterpolate.variogram.utils.exceptions import validate_plot_attributes_for_experimental_variogram_class
+
+
+class DirectionalVariogram:
+    """
+    Class prepares four directional variograms and one isotropic variogram.
+
+    Parameters
+    ----------
+    input_array : numpy array, list, tuple
+        * As a ``list`` and ``numpy array``: coordinates and their values: ``(pt x, pt y, value)``,
+        * as a ``dict``: ``polyset = {'points': numpy array with coordinates and their values}``,
+        * as a ``Blocks``: ``Blocks.polyset['points']``.
+
+    step_size : float
+        The distance between lags within each the points are included in the calculations.
+
+    max_range : float
+        The maximum range of analysis.
+
+    weights : numpy array, default=None
+        Weights assigned to the points, index of weight must be the same as index of point.
+
+    tolerance : float (in range [0, 1]), default=0.2
+        If ``tolerance`` is 0 then points must be placed at a single line with the beginning in the origin of
+        the coordinate system and the angle given by y axis and direction parameter. If ``tolerance`` is ``> 0`` then
+        the bin is selected as an elliptical area with major axis pointed in the same direction as the line
+        for 0 tolerance.
+
+        * The major axis size == ``step_size``.
+        * The minor axis size is ``tolerance * step_size``
+        * The baseline point is at a center of the ellipse.
+        * The ``tolerance == 1`` creates an omnidirectional semivariogram.
+
+    method : str, default = triangular
+        The method used for neighbors selection. Available methods:
+
+        * "triangle" or "t", default method where a point neighbors are selected from a triangular area,
+        * "ellipse" or "e", the most accurate method but also the slowest one.
+
+    Attributes
+    ----------
+    ds : numpy array
+        See the ``input_array`` parameter.
+
+    step_size : float
+        See the ``step_size`` parameter.
+
+    max_range : float
+        See the ``max_range`` parameter.
+
+    tolerance : float
+        See the ``tolerance`` parameter.
+
+    weights : float
+        See the ``weights`` parameter.
+
+    directions : Dict
+        Dictionary where keys are directions: NS, WE, NE-SW, NW-SE, and values are angles: -90, 0, -45, 45.
+
+    method : str, default = triangular
+        See the ``method`` parameter.
+
+    directional_variograms : Dict
+        Dictionary with five variograms:
+
+        * ``ISO``: isotropic,
+        * ``NS``: North-South axis,
+        * ``WE``: West-East axis,
+        * ``NE-SW``: Northeastern-Southwestern axis,
+        * ``NW-SE``: Northwestern-Southeastern axis.
+
+    Methods
+    -------
+    get()
+        Returns copy of calculated directional variograms.
+
+    show()
+        Plot all variograms on a single plot.
+    """
+
+    def __init__(self, input_array: np.array, step_size: float, max_range: float, weights=None, tolerance=0.2,
+                 method='t'):
+
+        self.ds = input_array
+        self.step_size = step_size
+        self.max_range = max_range
+        self.tolerance = tolerance
+        self.weights = weights
+        self.method = method
+
+        self.directions = {
+            'NS': -90,
+            'WE': 0,
+            'NE-SW': -45,
+            'NW-SE': 45
+        }
+
+        self.directional_variograms = {}
+
+        self._build_experimental_variograms()
+
+    def _build_experimental_variograms(self):
+
+        isotropic = build_experimental_variogram(self.ds, self.step_size, self.max_range, weights=self.weights)
+        self.directional_variograms['ISO'] = isotropic
+
+        for idx, val in self.directions.items():
+            variogram = build_experimental_variogram(self.ds,
+                                                     self.step_size,
+                                                     self.max_range,
+                                                     weights=self.weights,
+                                                     direction=val,
+                                                     tolerance=self.tolerance,
+                                                     method=self.method)
+            self.directional_variograms[idx] = variogram
+
+    def get(self):
+        return self.directional_variograms.copy()
+
+    def show(self):
+        if self.directional_variograms:
+            _lags = self.directional_variograms['ISO'].lags
+            _ns = self.directional_variograms['NS'].experimental_semivariances
+            _we = self.directional_variograms['WE'].experimental_semivariances
+            _nw_se = self.directional_variograms['NW-SE'].experimental_semivariances
+            _ne_sw = self.directional_variograms['NE-SW'].experimental_semivariances
+            _iso = self.directional_variograms['ISO'].experimental_semivariances
+
+            plt.figure(figsize=(20, 8))
+            plt.plot(_lags, _iso, color='#1b9e77')
+            plt.plot(_lags, _ns, '--', color='#d95f02')
+            plt.plot(_lags, _we, '--', color='#7570b3')
+            plt.plot(_lags, _nw_se, '--', color='#e7298a')
+            plt.plot(_lags, _ne_sw, '--', color='#66a61e')
+            plt.title('Comparison of experimental semivariance models')
+            plt.legend(['Isotropic',
+                        'NS',
+                        'WE',
+                        'NW-SE',
+                        'NE-SW'])
+            plt.xlabel('Distance')
+            plt.ylabel('Semivariance')
+            plt.show()
 
 
 class ExperimentalVariogram:
@@ -38,12 +181,12 @@ class ExperimentalVariogram:
         Weights assigned to points, index of weight must be the same as index of point.
 
     direction : float (in range [0, 360]), default=0
-        A direction of semivariogram, values from 0 to 360 degrees:
-
-        * 0 or 180: is NS direction,
-        * 90 or 270 is EW direction,
-        * 45 or 225 is NE-SW direction,
-        * 135 or 315 is NW-SE direction.
+        Direction of semivariogram, values from 0 to 360 degrees:
+        
+        - 0 or 180: is E-W,
+        - 90 or 270 is N-S,
+        - 45 or 225 is NW-SE,
+        - 135 or 315 is NE-SW.
 
     tolerance : float (in range [0, 1]), default=1
         If ``tolerance`` is 0 then points must be placed at a single line with the beginning in the origin of
@@ -55,6 +198,12 @@ class ExperimentalVariogram:
         * The minor axis size is ``tolerance * step_size``
         * The baseline point is at a center of the ellipse.
         * The ``tolerance == 1`` creates an omnidirectional semivariogram.
+
+    method : str, default = triangular
+        The method used for neighbors selection. Available methods:
+
+        * "triangle" or "t", default method where a point neighbors are selected from a triangular area,
+        * "ellipse" or "e", the most accurate method but also the slowest one.
 
     is_semivariance : bool, optional, default=True
         Should semivariance be calculated?
@@ -123,6 +272,9 @@ class ExperimentalVariogram:
     tol : float
         Derived from the ``tolerance`` parameter.
 
+    method : str
+        See the ``method`` parameter.
+
     Methods
     -------
     plot()
@@ -171,6 +323,7 @@ class ExperimentalVariogram:
                  weights=None,
                  direction: float = 0.0,
                  tolerance: float = 1.0,
+                 method='t',
                  is_semivariance=True,
                  is_covariance=True,
                  is_variance=True):
@@ -197,6 +350,7 @@ class ExperimentalVariogram:
         self.weights = weights
         self.direct = direction
         self.tol = tolerance
+        self.method = method
 
         self.__c_sem = is_semivariance
         self.__c_cov = is_covariance
@@ -240,7 +394,7 @@ class ExperimentalVariogram:
         -----
         AttributeSetToFalseWarning
             Warning invoked when plotting parameter for semivariance, covariance or variance is set to ``True`` but
-            class atrributes to calculate those indices are set to ``False``.
+            class attributes to calculate those indices are set to ``False``.
         """
 
         # Validate parameters
@@ -300,7 +454,8 @@ class ExperimentalVariogram:
             max_range=self.mx_rng,
             weights=self.weights,
             direction=self.direct,
-            tolerance=self.tol
+            tolerance=self.tol,
+            method=self.method
         )
 
     def __repr__(self):
@@ -376,7 +531,8 @@ def build_experimental_variogram(input_array: np.array,
                                  max_range: float,
                                  weights: np.array = None,
                                  direction: float = 0,
-                                 tolerance: float = 1) -> ExperimentalVariogram:
+                                 tolerance: float = 1,
+                                 method='t') -> ExperimentalVariogram:
     """
     Function prepares:
       - experimental semivariogram,
@@ -398,12 +554,12 @@ def build_experimental_variogram(input_array: np.array,
         Weights assigned to points, index of weight must be the same as index of point.
 
     direction : float (in range [0, 360]), default = 0
-        A direction of semivariogram, values from 0 to 360 degrees:
-
-        * 0 or 180: is NS direction,
-        * 90 or 270 is EW direction,
-        * 45 or 225 is NE-SW direction,
-        * 135 or 315 is NW-SE direction.
+        Direction of semivariogram, values from 0 to 360 degrees:
+        
+        - 0 or 180: is E-W,
+        - 90 or 270 is N-S,
+        - 45 or 225 is NW-SE,
+        - 135 or 315 is NE-SW.
 
     tolerance : float (in range [0, 1]), optional, default=1
         If ``tolerance`` is 0 then points must be placed at a single line with the beginning in the origin of
@@ -415,6 +571,12 @@ def build_experimental_variogram(input_array: np.array,
         * the minor axis size is ``tolerance * step_size``,
         * the baseline point is at a center of the ellipse,
         * the ``tolerance == 1`` creates an omnidirectional semivariogram.
+
+    method : str, default = triangular
+        The method used for neighbors selection. Available methods:
+
+        * "triangle" or "t", default method where a point neighbors are selected from a triangular area,
+        * "ellipse" or "e", the most accurate method but also the slowest one.
 
     Returns
     -------
@@ -469,6 +631,7 @@ def build_experimental_variogram(input_array: np.array,
         weights=weights,
         direction=direction,
         tolerance=tolerance,
+        method=method,
         is_semivariance=True,
         is_covariance=True,
         is_variance=True